backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pyodbc>=5.1.0
pandas>=2.2.0
openpyxl>=3.1.0
aiosqlite>=0.20.0
exchangelib>=5.2.0
msal>=1.28.0
anthropic>=0.26.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
chromadb>=0.5.0
pypdf>=4.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
Pillow>=10.3.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
rapidfuzz>=3.9.0
tqdm>=4.66.0
httpx>=0.27.0
python-multipart>=0.0.9
celery>=5.4.0
redis>=5.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
azure-storage-blob>=12.20.0
python-docx>=1.1.0
anthropic[bedrock]>=0.26.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt<4
email-validator>=2.1.0
